Charging behaviours

+++ src/main/java/Common/Messages/ChargeResponse.java
@@ -4,8 +4,8 @@ package Common.Messages;
  * Created by janbaraniewski on 03/05/16.
  */
 public class ChargeResponse extends Message {
-    int waitingTime;
-    public ChargeResponse(int waitingTime){
+    long waitingTime;
+    public ChargeResponse(long waitingTime){
         this.waitingTime = waitingTime;
     }
 }
